make offchain microservice to call api-gateway service only
